Moving shared/topic_tools to main
