+++ kalite/coachreports/api_views.py
@@ -21,10 +21,10 @@ from django.utils.translation import ugettext as _
 from .forms import DataForm
 from config.models import Settings
 from main.models import VideoLog, ExerciseLog, UserLog, UserLogSummary
+from main.topic_tools import get_topic_by_path, get_node_cache
 from securesync import facility_required
 from securesync.models import Facility, FacilityUser, FacilityGroup, DeviceZone, Device
 from settings import LOG as logging
-from shared.topic_tools import get_topic_by_path, get_node_cache
 from testing.decorators import allow_api_profiling
 from utils.internet import StatusException, JsonResponse, api_handle_error_with_json
