+++ kalite/coachreports/management/commands/generaterealdata.py
@@ -34,9 +34,9 @@ from django.db import transaction
 import settings
 import securesync
 from main.models import ExerciseLog, VideoLog, UserLog
+from main.topic_tools import get_topic_videos, get_topic_exercises
 from securesync.models import Facility, FacilityUser, FacilityGroup, Device, DeviceMetadata
 from settings import LOG as logging
-from shared.topic_tools import get_topic_videos, get_topic_exercises
 from utils.general import datediff
