+++ kalite/coachreports/views.py
@@ -18,11 +18,11 @@ from django.utils.translation import ugettext as _
 
 from .api_views import get_data_form, stats_dict
 from main.models import VideoLog, ExerciseLog, UserLog
+from main.topic_tools import get_topic_exercises, get_topic_videos, get_knowledgemap_topics, get_node_cache, get_topic_tree
 from securesync import facility_required
 from securesync.models import Facility, FacilityUser, FacilityGroup, DeviceZone, Device
 from settings import LOG as logging
 from shared.decorators import require_authorized_access_to_student_data, require_authorized_admin, get_user_from_request
-from shared.topic_tools import get_topic_exercises, get_topic_videos, get_knowledgemap_topics, get_node_cache, get_topic_tree
 from utils.general import max_none
 from utils.internet import StatusException
