+++ kalite/control_panel/views.py
@@ -24,11 +24,11 @@ except:
         pass
 from coachreports.views import student_view_context
 from main.models import ExerciseLog, VideoLog, UserLog, UserLogSummary
+from main.topic_tools import get_node_cache
 from securesync.forms import FacilityForm
 from securesync.models import Facility, FacilityUser, FacilityGroup, DeviceZone, Device, Zone, SyncSession
 from settings import LOG as logging
 from shared.decorators import require_authorized_admin, require_authorized_access_to_student_data
-from shared.topic_tools import get_node_cache
 from utils.internet import CsvResponse, render_to_csv
