+++ kalite/i18n/management/commands/verify_language_packs.py
@@ -19,8 +19,8 @@ from django.core.management.base import BaseCommand, CommandError
 
 import settings
 from i18n import lcode_to_django
+from main.topic_tools import get_node_cache
 from settings import LOG as logging
-from shared.topic_tools import get_node_cache
 from update_language_packs import handle_po_compile_errors
 from update_po import compile_po_files
 from utils.general import ensure_dir, datediff
