+++ kalite/khanload/api_views.py
@@ -38,10 +38,10 @@ from django.views.decorators.csrf import csrf_exempt
 
 import settings
 from main.models import ExerciseLog, VideoLog
+from main.topic_tools import get_node_cache
 from settings import LOG as logging
 from securesync.models import FacilityUser
 from shared.decorators import require_login
-from shared.topic_tools import get_node_cache
 from testing.asserts import central_server_only, distributed_server_only
 from utils.internet import JsonResponse, JsonResponseMessageError
