+++ kalite/main/api_views.py
@@ -26,12 +26,12 @@ from . import topicdata
 from .api_forms import ExerciseLogForm, VideoLogForm, DateTimeForm
 from .caching import backend_cache_page
 from .models import VideoLog, ExerciseLog
+from .topic_tools import get_flat_topic_tree
 from chronograph import force_job, job_status
 from config.models import Settings
 from i18n import lcode_to_ietf
 from securesync.models import FacilityGroup, FacilityUser
 from shared.decorators import require_admin
-from shared.topic_tools import get_flat_topic_tree
 from testing.decorators import allow_api_profiling
 from utils.general import break_into_chunks
 from utils.internet import api_handle_error_with_json, JsonResponse, JsonResponseMessage, JsonResponseMessageError, JsonResponseMessageWarning
