+++ kalite/main/tests/browser_tests.py
@@ -17,10 +17,10 @@ from django.utils.translation import ugettext as _
 
 import settings
 from main.models import ExerciseLog
+from main.topic_tools import get_exercise_paths
 from main.topicdata import NODE_CACHE
 from securesync.models import Facility, FacilityGroup, FacilityUser
 from settings import LOG as logging
-from shared.topic_tools import get_exercise_paths
 from testing.browser import BrowserTestCase
 from testing.decorators import distributed_server_test
 from utils.django_utils import call_command_with_output
