+++ kalite/main/tests/cache_tests.py
@@ -11,7 +11,7 @@ from django.utils import unittest
 
 import settings
 from main import caching
-from shared.topic_tools import get_node_cache
+from main.topic_tools import get_node_cache
 from testing.base import KALiteTestCase
 from testing.decorators import distributed_server_test
 from utils.django_utils import call_command_with_output
