+++ kalite/main/topicdata.py
@@ -1,4 +1,4 @@
-from shared import topic_tools
+import topic_tools
 
 TOPICS          = topic_tools.get_topic_tree()
 NODE_CACHE      = topic_tools.get_node_cache()
