+++ kalite/updates/views.py
@@ -27,9 +27,9 @@ from chronograph import force_job
 from config.models import Settings
 from control_panel.views import local_device_context
 from i18n import lcode_to_ietf, get_installed_language_packs, lang_best_name
+from main import topic_tools
 from securesync.models import Facility, FacilityUser, FacilityGroup, Device
 from securesync.views import require_admin
-from shared import topic_tools
 from shared.decorators import require_admin
 from utils.internet import am_i_online, JsonResponse
